new file:   .gitignore 	new file:   .idea/.gitignore 	new file:   .idea/Module_D.iml 	new file:   .idea/dbnavigator.xml 	new file:   .idea/inspectionProfiles/profiles_settings.xml 	new file:   .idea/misc.xml 	new file:   .idea/modules.xml 	new file:   .idea/vcs.xml 	modified:   NewsPaper/__pycache__/settings.cpython-310.pyc 	modified:   NewsPaper/settings.py 	modified:   db.sqlite3 	modified:   news/__pycache__/forms.cpython-310.pyc 	modified:   news/__pycache__/models.cpython-310.pyc 	modified:   news/__pycache__/views.cpython-310.pyc 	modified:   news/forms.py 	new file:   news/migrations/0002_alter_category_options_and_more.py 	modified:   news/models.py 	modified:   news/views.py 	new file:   requirements.txt 	modified:   sign/forms.py 	modified:   sign/views.py 	renamed:    templates/login.html -> templates/sign/login.html
